+++ packages/vscode-extension/src/commandHandlers/connection.ts
@@ -1,4 +1,4 @@
-import { ConnnectionResult } from '@neo4j-cypher/schema-poller';
+import { ConnnectionResult } from '@neo4j-cypher/query-tools';
 import { commands, Selection, TextEditor, window, workspace } from 'vscode';
 import {
   Connection,
@@ -196,8 +196,8 @@ export async function runCypher(): Promise<void> {
     const activeConnection = getActiveConnection();
 
     if (!activeConnection) {
-      await window.showErrorMessage(
-        `You need to be connected to Neo4j to run queries`,
+      void window.showErrorMessage(
+        CONSTANTS.MESSAGES.ERROR_DISCONNECTED_EXECUTION,
       );
 
       return;
@@ -226,9 +226,10 @@ export async function cypherFileFromSelection(): Promise<void> {
 
 export async function forceDisconnect(): Promise<void> {
   const activeConnection = getActiveConnection();
-  const { result, connection } =
-    await toggleConnectionAndUpdateDatabaseConnection(activeConnection);
-  displayMessageForConnectionResult(connection, result);
+
+  await toggleConnectionAndUpdateDatabaseConnection(activeConnection);
+
+  void window.showInformationMessage(CONSTANTS.MESSAGES.DISCONNECTED_MESSAGE);
 }
 
 export async function forceConnect(i: number): Promise<void> {