+++ packages/language-server/CHANGELOG.md
@@ -4,8 +4,11 @@
 
 ### Patch Changes
 
+- 5bfdac2: restructure packages as part of move to pnpm
 - Updated dependencies [88e4930]
+- Updated dependencies [5bfdac2]
   - @neo4j-cypher/language-support@2.0.0-next.19
+  - @neo4j-cypher/query-tools@2.0.0-next.19
 
 ## 2.0.0-next.19
 
@@ -178,4 +181,4 @@
 
 - Updated dependencies [5819f6385]
   - @neo4j-cypher/language-support@2.0.0-next.0
-  - @neo4j-cypher/schema-poller@2.0.0-next.0
+  - @neo4j-cypher/query-tools@2.0.0-next.0