+++ packages/language-support/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Patch Changes
 
 - 88e4930: Fixes backticking of parameters and dbNames/aliases
+- 5bfdac2: restructure packages as part of move to pnpm
 
 ## 2.0.0-next.18