+++ packages/vscode-extension/CHANGELOG.md
@@ -1,5 +1,10 @@
 # neo4j-for-vscode
 
+## 1.10.0
+
+- Adds new version of the Cypher formatting
+- Asks the user to set missing parameters when trying to execute a query
+
 ## 1.9.0
 
 - Adds parameter pane to VSCode.